requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
youtube-transcript-api==0.6.1
pytube==15.0.0
openai>=1.55.3
python-dotenv==1.0.0
requests==2.31.0